+++ src/Repository/UserRepository.php
@@ -6,15 +6,11 @@ namespace App\Repository;
 use App\AppConstants;
 use App\Entity\User;
 use App\Entity\UserRoles;
-use App\Resource\UsersStatsOutput;
 use App\Traits\ToolsTrait;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\ORM\NonUniqueResultException;
-use Doctrine\ORM\NoResultException;
 use Doctrine\ORM\Query\Expr\Join;
 use Doctrine\ORM\QueryBuilder;
 use Doctrine\Persistence\ManagerRegistry;
-use Psr\Log\LoggerInterface;
 
 /**
  * @method User|null find($id, $lockMode = null, $lockVersion = null)
@@ -28,85 +24,14 @@ class UserRepository extends ServiceEntityRepository
 {
     use ToolsTrait;
 
-    public const AVAILABLE_FILTERS = [AppConstants::WITH_DETAILS];
     public const USER_ALIAS = 'u';
 
-    private LoggerInterface $logger;
-
-    public function __construct(ManagerRegistry $registry, LoggerInterface $logger)
+    public function __construct(ManagerRegistry $registry)
     {
         parent::__construct($registry, User::class);
-        $this->logger = $logger;
+
     }
 
-    public function getUserStats(array $filters): UsersStatsOutput
-    {
-        $rvId = null;
-
-        if (array_key_exists('rvid', $filters)) {
-            $rvId = (int)$filters['rvid'];
-        }
-
-        $uid = array_key_exists('uid', $filters) ? (int)$filters['uid'] : null;
-        $role = (array_key_exists('role', $filters) && !empty($filters['role'])) ? $filters['role'] : null;
-        $registrationYear = array_key_exists('registrationDate', $filters) ? (int)$filters['registrationDate'] : null;
-        $withDetails = array_key_exists(AppConstants::WITH_DETAILS, $filters);
-
-        $statResource = new UsersStatsOutput();
-        $statResource->setAvailableFilters(self::AVAILABLE_FILTERS);
-        $statResource->setRequestedFilters($filters);
-        $statResource->setName('nbUsers');
-
-        $userStatsQuery = $this->findByReviewQuery($rvId, $uid, $role, $withDetails, $registrationYear)->getQuery();
-        $userStats = $userStatsQuery->getArrayResult();
-
-
-        try {
-            $nbUsers = (int)$this->findByReviewQuery($rvId, $uid, $role, false, $registrationYear)->getQuery()->getSingleScalarResult();
-        } catch (NoResultException | NonUniqueResultException $e) {
-            $nbUsers = null;
-            $this->logger->error($e->getMessage());
-        }
-
-        $details = null;
-
-        if ($withDetails) {
-
-            if ($rvId && !$role) {
-
-                $rvIdResult = $this->applyFilterBy($userStats, 'rvid', (string)$rvId);
-
-                if (array_key_exists($rvId, $rvIdResult)) {
-                    $details = $this->reformatData($rvIdResult[$rvId]);
-                    $statResource->setDetails($details);
-                }
-
-            } elseif (!$rvId && $role) {
-                $roleResult = $this->applyFilterBy($userStats, 'role', $role);
-                $statResource->setDetails($roleResult);
-            } elseif ($rvId && $role) {
-
-                $details = $this->applyFilterBy($userStats, 'rvid', (string)$rvId);
-
-                if (array_key_exists($rvId, $details)) {
-                    $details = $this->reformatData($details[$rvId]);
-                }
-
-                $roleResult = array_key_exists($role, $details) ? $details[$role] : [];
-                $statResource->setDetails($roleResult);
-
-            }
-
-            $details = array_key_exists($rvId, $this->reformatData($userStats)) ?
-                $this->reformatData($userStats)[$rvId] :
-                $this->reformatData($userStats) ;
-        }
-
-        $statResource->setValue($nbUsers);
-        $statResource->setDetails($details);
-
-        return $statResource;
-    }
 
     /**
      * get users by review query
@@ -119,8 +44,7 @@ class UserRepository extends ServiceEntityRepository
      */
     public function findByReviewQuery(int $rvId = null, $uid = null, string $role = null, bool $withDetails = false, int $registrationYear = null): QueryBuilder
     {
-        $userRolesRepo = $this->getEntityManager()->getRepository(UserRoles::class);
-        return $userRolesRepo->getUserRolesStatsQuery($rvId, $uid, $role, $withDetails);
+        return $this->getEntityManager()->getRepository(UserRoles::class)->getUserRolesStatsQuery($rvId, $uid, $role, $withDetails);
     }
 
     /**
@@ -190,43 +114,5 @@ class UserRepository extends ServiceEntityRepository
         $qb->addGroupBy("$userRolesAlias.roleid");
         $qb->addGroupBy("$userAlias1.uid");
         return $qb;
-    }
-
-    private function reformatData(array $array): array
-    {
-        $rvId = null;
-        $role = null;
-        $nbUsers = 0;
-        $result = [];
-
-        foreach ($array as $key => $value) {
-
-            foreach ($value as $k => $v) {
-
-                if ($k === 'rvid' && $v !== $rvId) {
-                    $rvId = $v;
-                }
-
-                if ($k === 'role' && $v !== $role) {
-                    $role = $v;
-                    $nbUsers = 0;
-                }
-
-                if ($k === 'nbUsers') {
-                    $nbUsers += $v;
-                }
-
-
-                if ($rvId === null && null !== $role) {
-                    $result[$role]['nbUsers'] = $nbUsers;
-
-                } elseif(null !== $role) {
-                    $result[$rvId][$role]['nbUsers'] = $nbUsers;
-
-                }
-
-            }
-        }
-        return $result;
     }
 }