+++ src/Resource/AbstractStatResource.php
@@ -86,7 +86,7 @@ abstract class AbstractStatResource
     /**
      * @return float|array|null
      */
-    public function getValue(): float|array|null
+    public function getValue(): int|float|array|null
     {
         return $this->value;
     }
@@ -95,7 +95,7 @@ abstract class AbstractStatResource
      * @param float|array|null $value
      * @return $this
      */
-    public function setValue(float|array|null $value): self
+    public function setValue(int|float|array|null $value): self
     {
         $this->value = $value;
         return $this;