+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade from Symfony 6.4 to 7.0
 - Adding sections End points for sections
 - Adding Extra Data: statusLabel and repository
+- refactoring + adding measurement's unit to publication delay indicators.
 ## v1.1.6 2025-02-14
 ### Fixed
 - Improvements: related to the option to ignore statistics before a given date